requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
openpyxl==3.1.5
prophet==1.1.5
scikit-learn==1.5.2
google-genai==1.19.0
plotly==5.24.1
streamlit==1.39.0
fpdf2==2.8.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.9
httpx>=0.28.1
numpy==1.26.4
scipy==1.14.1